+++ pkg/grpc/messaging_service/server.go
@@ -2,11 +2,11 @@ package messaging_service
 
 import (
 	"context"
-	"log"
 	"net"
 	"os"
 	"os/signal"
 
+	"github.com/coneno/logger"
 	api "github.com/influenzanet/messaging-service/pkg/api/messaging_service"
 	"github.com/influenzanet/messaging-service/pkg/dbs/messagedb"
 	"github.com/influenzanet/messaging-service/pkg/types"
@@ -43,7 +43,7 @@ func RunServer(ctx context.Context, port string,
 ) error {
 	lis, err := net.Listen("tcp", ":"+port)
 	if err != nil {
-		log.Fatalf("failed to listen: %v", err)
+		logger.Error.Fatalf("failed to listen: %v", err)
 	}
 
 	// register service
@@ -59,14 +59,14 @@ func RunServer(ctx context.Context, port string,
 	go func() {
 		for range c {
 			// sig is a ^C, handle it
-			log.Println("shutting down gRPC server...")
+			logger.Info.Println("shutting down gRPC server...")
 			server.GracefulStop()
 			<-ctx.Done()
 		}
 	}()
 
 	// start gRPC server
-	log.Println("starting gRPC server...")
-	log.Println("wait connections on port " + port)
+	logger.Info.Println("starting gRPC server...")
+	logger.Info.Println("wait connections on port " + port)
 	return server.Serve(lis)
 }